// projects/summarization-api-demo/src/app/ai/enums/capabilities-core-options.enum.ts
export enum AISummarizerType { 
    TLDR = 'tl;dr', 
    KEYPOINTS = 'key-points', 
    TEASER = 'teaser', 
    HEADLINE = 'headline' 
};

export enum AISummarizerFormat { 
    PLAIN_TEXT = 'plain-text', 
    MARKDOWN = 'markdown' 
};

export enum AISummarizerLength { 
    SHORT = 'short', 
    MEDIUM = 'medium', 
    LONG = 'long' 
};